requirements.txt
streamlit
python-dotenv
streamlit-extras
google.generative
googlesearch-python 
beautifulsoup4
python-pptx
